Provisional change to workspaces paths (#3).
