+++ lerna.json
@@ -1,6 +1,6 @@
 {
   "npmClient": "yarn",
   "useWorkspaces": true,
-  "version": "0",
-  "packages": ["lib/*", "examples/**/*"]
+  "version": "independent",
+  "packages": ["packages/*"]
 }
