+++ package.json
@@ -2,8 +2,7 @@
   "name": "root",
   "private": true,
   "workspaces": [
-    "lib/*",
-    "examples/**/*"
+    "packages/*"
   ],
   "scripts": {
     "build": "lerna run build",
@@ -11,7 +10,7 @@
     "test": "lerna run test --"
   },
   "devDependencies": {
-    "lerna": "^8.1.8",
+    "lerna": "^8.1.9",
     "lerna-update-wizard": "^1.1.2"
   },
   "resolutions": {
