+++ packages/examples/cli_client_gl/vite.config.ts
