+++ packages/examples/cli_server_metal/package.json
