+++ packages/examples/cli_server_metal/src/index.ts
