+++ packages/examples/cli_server_metal/vite.config.ts
