+++ packages/lib/binding.gyp
