+++ packages/lib/package.json
@@ -14,12 +14,17 @@
   },
   "main": "dist/node-syphon.js",
   "module": "dist/node-syphon.mjs",
-  "types": "dist/types/main/index.d.ts",
+  "types": "dist/main/index.d.ts",
   "exports": {
     ".": {
       "import": "./dist/node-syphon.mjs",
       "require": "./dist/node-syphon.js",
-      "types": "./dist/types/main/index.d.ts"
+      "types": "./dist/main/index.d.ts"
+    },
+    "./universal": {
+      "import": "./dist/node-syphon-universal.mjs",
+      "require": "./dist/node-syphon-universal.js",
+      "types": "./dist/main/universal.d.ts"
     }
   },
   "files": [
@@ -32,7 +37,7 @@
     "build": "yarn build:syphon && node-gyp configure && node-gyp build && yarn build:main && yarn build:directory && rimraf build",
     "build:syphon": "./scripts/build-syphon.sh Release",
     "build:addon": "node-gyp configure && node-gyp build && mv build/Release/syphon.node dist/bin/syphon.node && rimraf build",
-    "build:main": "vite build --config vite.config.main.mts",
+    "build:main": "vite build --config vite.config.main.mts && vite build --config vite.config.universal.mts",
     "build:directory": "vite build --config vite.config.directory.mts"
   },
   "devDependencies": {
