+++ packages/lib/src/common/addon.ts
