+++ packages/lib/src/common/types.ts
@@ -1,0 +1,10 @@
+export const SyphonServerDescriptionNameKey = 'SyphonServerDescriptionNameKey';
+export const SyphonServerDescriptionAppNameKey = 'SyphonServerDescriptionAppNameKey';
+export const SyphonServerDescriptionUUIDKey = 'SyphonServerDescriptionUUIDKey';
+
+export type SyphonServerDescriptionPropertyKey =
+  | typeof SyphonServerDescriptionNameKey
+  | typeof SyphonServerDescriptionAppNameKey
+  | typeof SyphonServerDescriptionUUIDKey;
+
+export type SyphonServerDescription = Record<SyphonServerDescriptionPropertyKey, string>;
