+++ packages/lib/src/obj-cpp/classes/NodeSyphonFramePromiseWorker.h
