+++ packages/lib/src/obj-cpp/classes/NodeSyphonMetalServer.h
