+++ packages/lib/src/obj-cpp/classes/NodeSyphonOpenGLClient.mm
