+++ packages/lib/src/obj-cpp/classes/NodeSyphonServerDirectory.h
