+++ packages/lib/src/obj-cpp/classes/NodeSyphonServerDirectory.mm
