+++ packages/lib/src/obj-cpp/helpers/NodeSyphonHelpers.mm
