+++ packages/lib/src/obj-cpp/helpers/macros.h
