+++ packages/lib/src/obj-cpp/main.h
