+++ packages/lib/vite.config.directory.mts
