+++ packages/lib/vite.config.main.mts
